Lock thread safe storage with RwLock
We would like for multiple thread sto be able to read from the same
thread safe storage simultaneously, as long as they don't attempt to
write.

The Mutex type does not support this behaviour. Instead the RwLock type
allows multipe readers and up to one writer. The downside is that the
types locked behind an RwLock need to also be Sync. Since we will use
arithmetic data types in our matrices, this restriction is irrelevant to
us.

Closes #12.